+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/MainConfiguration.php
@@ -41,8 +41,8 @@ class MainConfiguration implements ConfigurationInterface
      */
     public function getConfigTreeBuilder()
     {
-        $tb = new TreeBuilder();
-        $rootNode = $tb->root('security');
+        $tb = new TreeBuilder('security');
+        $rootNode = $tb->getRootNode();
 
         $rootNode
             ->beforeNormalization()